Observable error identified
When incoming velocity switches from positive to negative, position spikes massively instead of subtracting as it should. Negative numbers were being converted into very large positive numbers due to a failure with the sign extension. Fixed by changing a 32 to a 33 in the sign extender module.

Next issue: Border control is not functioning correctly.

Solution: It is a parameter issue, it seems the influence of the turnfactor is getting washed out by our method of proportional speed control.